requirements.txt
# Core FastAPI dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# MongoDB integration
motor>=3.0.0
pymongo>=4.0.0

# PDF processing
PyMuPDF>=1.23.0
pdfminer.six>=20221105

# HTTP client for OpenRouter API
httpx>=0.24.0

# Data validation and serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
typing-extensions>=4.5.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: For better JSON handling
orjson>=3.8.0
